Allow hashie 4
Tests all pass, is that enough to be sure that hashie 4 is just fine? Gemspec still allows hashie 3 too, it doesn't force an app to update. It just allows an app to use hashie 4 instead.

I have an old gem that's still using confstruct, and I didn't wnat using my gem to force an app to use hashie 3, that might be incompatible with other dependencies. That's the motivation here. My gem does have it's own tests pass against a confstruct using hashie 4.

If we merge this, I'd like to do a confstruct release. Since it doesn't force an upgrade to hashie 4, probably does not need to be a major version release, just minor?